modify :: adapter
