+++ application-infrastructure/src/main/kotlin/hs/kr/equus/application/domain/admin/presentation/WebAdminAdapter.kt
@@ -12,6 +12,8 @@ import hs.kr.equus.application.domain.application.usecase.PrintApplicantCodesUse
 import hs.kr.equus.application.global.excel.generator.PrintApplicantCodesGenerator
 import hs.kr.equus.application.domain.application.usecase.QueryStaticsCountUseCase
 import hs.kr.equus.application.domain.application.usecase.dto.response.GetStaticsCountResponse
+import hs.kr.equus.application.domain.score.presentation.dto.reponse.GetScoreResponse
+import hs.kr.equus.application.domain.score.usecase.QueryStaticsScoreUseCase
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.ModelAttribute
 import org.springframework.web.bind.annotation.PathVariable
@@ -27,8 +29,27 @@ class WebAdminAdapter(
     private val getApplicationUseCase: GetApplicationUseCase,
     private val getApplicantsUseCase: GetApplicantsUseCase,
     private val printApplicantCodesUseCase: PrintApplicantCodesUseCase,
-    private val queryStaticsCountUseCase: QueryStaticsCountUseCase
+    private val queryStaticsCountUseCase: QueryStaticsCountUseCase,
+    private val queryStaticsScoreUseCase: QueryStaticsScoreUseCase
 ) {
+
+    @GetMapping("/statics/score")
+    fun queryStaticsScore(): List<GetScoreResponse> =
+        queryStaticsScoreUseCase.execute().map {
+            GetScoreResponse(
+                applicationType = it.applicationType,
+                isDaejeon = it.isDaejeon,
+                score_79 = it.score_79,
+                score80_92 = it.score80_92,
+                score93_105 = it.score93_105,
+                score106_118 = it.score106_118,
+                score119_131 = it.score119_131,
+                score132_144 = it.score132_144,
+                score145_157 = it.score145_157,
+                score158_170 = it.score158_170
+            )
+        }
+
     @GetMapping("/statics/count")
     fun queryStaticsCount(): List<GetStaticsCountResponse> =
         queryStaticsCountUseCase.execute()
