+++ application-infrastructure/src/main/kotlin/hs/kr/equus/application/domain/score/presentation/WebScoreAdapter.kt
@@ -1,7 +1,5 @@
 package hs.kr.equus.application.domain.score.presentation
 
-import hs.kr.equus.application.domain.score.presentation.dto.reponse.GetScoreResponse
-import hs.kr.equus.application.domain.score.usecase.QueryStaticsScoreUseCase
 import hs.kr.equus.application.domain.score.usecase.QueryMyTotalScoreUseCase
 import hs.kr.equus.application.domain.score.usecase.dto.response.QueryTotalScoreResponse
 import org.springframework.web.bind.annotation.GetMapping
@@ -11,26 +9,8 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 @RequestMapping("/score")
 class WebScoreAdapter(
-    private val queryStaticsScoreUseCase: QueryStaticsScoreUseCase,
     private val queryMyTotalScoreUseCase: QueryMyTotalScoreUseCase
 ) {
-    @GetMapping
-    fun queryStaticsScore(): List<GetScoreResponse> =
-        queryStaticsScoreUseCase.execute().map {
-            GetScoreResponse(
-                applicationType = it.applicationType,
-                isDaejeon = it.isDaejeon,
-                score_79 = it.score_79,
-                score80_92 = it.score80_92,
-                score93_105 = it.score93_105,
-                score106_118 = it.score106_118,
-                score119_131 = it.score119_131,
-                score132_144 = it.score132_144,
-                score145_157 = it.score145_157,
-                score158_170 = it.score158_170
-            )
-        }
-    
     @GetMapping("/total-score")
     fun queryMyTotalScore(): QueryTotalScoreResponse? =
         queryMyTotalScoreUseCase.execute()
